+++ tests/integration-network/src/mockchain/mod.rs
@@ -7,3 +7,4 @@ mod notes;
 mod sibling;
 mod support;
 mod swapp;
+mod tx_script_args;